+++ JotunnLib/Managers/PieceManager.cs
@@ -2,7 +2,6 @@
 using System;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 
 namespace JotunnLib.Managers
 {
