+++ TestMod/Prefabs/TestCubePrefab.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using UnityEngine;
-using JotunnLib.Entities;
+using JotunnLib.Configs;
 
 namespace TestMod.Prefabs
 {
